Maintain Float64 Points

+++ src/analysis/feasible_paths.jl
@@ -1,17 +1,17 @@
 
 """
-    get_feasible_matrix(waypoints::Vector{Point{2, Float32}}, exclusions::DataFrame)
+    get_feasible_matrix(waypoints::Vector{Point{2, Float64}}, exclusions::DataFrame)
 
 Create a distance matrix between waypoints accounting for environmental constraints.
 
 # Arguments
-- `waypoint::Vector{Point{2, Float32}s` : Vector of lat long tuples.
+- `waypoint::Vector{Point{2, Float64}s` : Vector of lat long tuples.
 - `exclusions::DataFrame` : DataFrame containing exclusion zones representing given vehicle's cumulative environmental constraints.
 
 # Returns
 Feasible distance matrix between waypoints.
 """
-function get_feasible_matrix(waypoints::Vector{Point{2, Float32}}, exclusions::DataFrame)::Matrix{Float64}
+function get_feasible_matrix(waypoints::Vector{Point{2, Float64}}, exclusions::DataFrame)::Matrix{Float64}
     n_waypoints = length(waypoints) - 1
     feasible_matrix = zeros(Float64, n_waypoints, n_waypoints)
 
@@ -26,20 +26,20 @@ function get_feasible_matrix(waypoints::Vector{Point{2, Float32}}, exclusions::D
 end
 
 """
-    shortest_feasible_path(line_pts::Tuple{Point{2, Float32}, Point{2, Float32}}, exclusions::DataFrame)
+    shortest_feasible_path(line_pts::Tuple{Point{2, Float64}, Point{2, Float64}}, exclusions::DataFrame)
 
 Find the shortest feasible path between two points.
 Use A* between all vertices on polygons that intersect with straight line to finish, from start pt and any other intersecting polygons.
 
 # Arguments
-- `line_pts::Tuple{Point{2, Float32}, Point{2, Float32}}`: A tuple containing two points on a line.
+- `line_pts::Tuple{Point{2, Float64}, Point{2, Float64}}`: A tuple containing two points on a line.
 - `exclusions::DataFrame`: A DataFrame containing exclusions.
 
 # Returns
 - `dist::Float64`: The distance of the shortest feasible path.
 - `path::Vector{SimpleWeightedGraph{Int64, Float64}.Edge}`: The shortest feasible path as a vector of edges.
 """
-function shortest_feasible_path(line_pts::Tuple{Point{2, Float32}, Point{2, Float32}}, exclusions::DataFrame)
+function shortest_feasible_path(line_pts::Tuple{Point{2, Float64}, Point{2, Float64}}, exclusions::DataFrame)
     pts = Set([line_pts[1]])
 
     new_pts = extract_unique_vertices(line_pts, exclusions)
@@ -63,19 +63,19 @@ function shortest_feasible_path(line_pts::Tuple{Point{2, Float32}, Point{2, Floa
 end
 
 """
-    extract_unique_vertices(exclusions::DataFrame)::Vector{Point{2, Float32}}
+    extract_unique_vertices(exclusions::DataFrame)::Vector{Point{2, Float64}}
 
 Extracts unique vertices from the given DataFrame of exclusions.
 
 # Arguments
-- `line_pts::Tuple{Point{2, Float32}, Point{2, Float32}}`: A tuple containing start/emd points of a line.
+- `line_pts::Tuple{Point{2, Float64}, Point{2, Float64}}`: A tuple containing start/emd points of a line.
 - `exclusions::DataFrame`: The DataFrame containing exclusions.
 
 # Returns
-- `Vector{Point{2, Float32}}`: A vector of unique vertices.
+- `Vector{Point{2, Float64}}`: A vector of unique vertices.
 """
-function extract_unique_vertices(line_pts::Tuple{Point{2, Float32}, Point{2, Float32}}, exclusions::DataFrame)::Union{Nothing, Vector{Point{2, Float32}}}
-    unique_vertices = Set{Point{2, Float32}}()
+function extract_unique_vertices(line_pts::Tuple{Point{2, Float64}, Point{2, Float64}}, exclusions::DataFrame)::Union{Nothing, Vector{Point{2, Float64}}}
+    unique_vertices = Set{Point{2, Float64}}()
 
     polygons = intersecting_polygons(line_pts, exclusions)
 
@@ -86,7 +86,7 @@ function extract_unique_vertices(line_pts::Tuple{Point{2, Float32}, Point{2, Flo
 
         for i in 0:n_pts - 1
             x, y, _ = AG.getpoint(exterior_ring, i)
-            point = Point{2, Float32}(x, y)
+            point = Point{2, Float64}(x, y)
 
             if !(point in unique_vertices)
                 push!(unique_vertices, point)
@@ -102,7 +102,7 @@ function extract_unique_vertices(line_pts::Tuple{Point{2, Float32}, Point{2, Flo
 end
 
 """
-    intersecting_polygons(pt_a::Point{2, Float32}, pt_b::Point{2, Float32}, exclusions::DataFrame)
+    intersecting_polygons(pt_a::Point{2, Float64}, pt_b::Point{2, Float64}, exclusions::DataFrame)
 
 Find polygons that intersect with a line segment.
 
@@ -113,7 +113,7 @@ Find polygons that intersect with a line segment.
 # Returns
 - `crossed_polygons`: A list of polygons that intersect with the line segment.
 """
-function intersecting_polygons(line_pts::Tuple{Point{2, Float32}, Point{2, Float32}}, exclusions::DataFrame)
+function intersecting_polygons(line_pts::Tuple{Point{2, Float64}, Point{2, Float64}}, exclusions::DataFrame)
     crossed_polygons = []
 
     line = AG.createlinestring([[line_pts[1][1], line_pts[1][2]], [line_pts[2][1], line_pts[2][2]]])
@@ -128,18 +128,18 @@ function intersecting_polygons(line_pts::Tuple{Point{2, Float32}, Point{2, Float
 end
 
 """
-    build_graph(pts::Vector{Point{2, Float32}}, exclusions::DataFrame)::SimpleWeightedGraph{Int64, Float64}
+    build_graph(pts::Vector{Point{2, Float64}}, exclusions::DataFrame)::SimpleWeightedGraph{Int64, Float64}
 
 Construct a simple weighted graph between given points that do not intersect exclusions.
 
 # Arguments
-- `pts::Vector{Point{2, Float32}`: A vector of 2D points.
+- `pts::Vector{Point{2, Float64}`: A vector of 2D points.
 - `exclusions::DataFrame`: A DataFrame containing exclusions.
 
 # Returns
 Simple weighted graph with distances between points.
 """
-function build_graph(pts::Vector{Point{2, Float32}}, exclusions::DataFrame)::SimpleWeightedGraph{Int64, Float64}
+function build_graph(pts::Vector{Point{2, Float64}}, exclusions::DataFrame)::SimpleWeightedGraph{Int64, Float64}
     g = SimpleWeightedGraph(length(pts))
     local_graphs = [SimpleWeightedGraph{Int64, Float64}(length(pts)) for _ in 1:nthreads()]
 
@@ -163,19 +163,19 @@ function build_graph(pts::Vector{Point{2, Float32}}, exclusions::DataFrame)::Sim
 end
 
 """
-    intersects_polygon(line::LineString{2, Float32}, exclusions::DataFrame)::Bool
+    intersects_polygon(line::LineString{2, Float64}, exclusions::DataFrame)::Bool
 
 Check if a line intersects with polygons in a dataframe.
 
 # Arguments
-- `pt_a::Point{2, Float32}`: The start point of the line.
-- `pt_b::Point{2, Float32}`: The end point of the line.
+- `pt_a::Point{2, Float64}`: The start point of the line.
+- `pt_b::Point{2, Float64}`: The end point of the line.
 - `exclusions::DataFrame`: The dataframe containing the polygon exclusions.
 
 # Returns
 - `Bool`: `true` if the line intersects with any polygon in the dataframe, `false` otherwise.
 """
-function intersects_polygon(pt_a::Point{2, Float32}, pt_b::Point{2, Float32}, exclusions::DataFrame)::Bool
+function intersects_polygon(pt_a::Point{2, Float64}, pt_b::Point{2, Float64}, exclusions::DataFrame)::Bool
     # TODO: Parallelise loop?
     for row in eachrow(exclusions)
         if AG.crosses(AG.createlinestring([[pt_a[1], pt_a[2]], [pt_b[1], pt_b[2]]]), row.geometry)
